### 2020-07-30 #2530 @onurvarol
Dear @kthyng thank you very much starting the process for review at JOSS. As you suggested we went through the list of reviewers and tried to identify potential reviewers with the relevant expertise and interest. You can see their usernames below.

felixhenninger, davekinkead, Andytwoods, mbod, adammichaelwood, tresoldi

Thank you very much and we are excited to experience JOSS review process.
