### 2020-08-04 #2530 @danielskatz
a different type of ping
