### 2020-08-14 #2530 @lorenanicole
@onurvarol wow what a collection of potential reviewers you have collected [here](https://bit.ly/joss-reviewers). As we need two, any you have particular recommendations on?
