### 2020-09-29 #2530 @lorenanicole
@emilydolson & @ajoer your names were provided as a possible reviewer for this submission for JOSS. Would you be interested in reviewing?
