### 2020-10-02 #2530 @lorenanicole
> @emilydolson & @ajoer your names were provided as a possible reviewer for this submission for JOSS. Would you be interested in reviewing?
>
> We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Hi. Just bumping this! Thanks!
