### 2020-10-07 #2530 @lorenanicole
@adammichaelwood & @davekinkead would you be willing to review this submission for JOSS?
