### 2020-10-08 #2530 @lorenanicole
@felixhenninger, we're looking for one more editor for this JOSS submission! Would you be willing to review this?
