### 2020-10-14 #2530 @sara-02
Hey @lorenanicole and @onurvarol I would be happy to review this package. I and my colleagures are actively working in the area of Twitter data collection and manual annotations, and would love to check out this package.
I just wrapped 2 of my pending reviews and have room for more.
Availablity: Fom 23rd Oct onwards. Let me know if it works for you guys.
