### 2020-10-25 #2780 @whedon
[REVIEW]: Social-Annotate: Browser extension to annotate and collect social media data
**Submitting author:** @onurvarol (<a href="http://orcid.org/0000-0002-3994-6106">Onur Varol</a>)
**Repository:** <a href="https://github.com/uluturki/social-annotate" target ="_blank">https://github.com/uluturki/social-annotate</a>
**Version:** v1.0.0
