### 2020-10-26 uluturki/social-annotate#19 @davekinkead
Clarify scholarly effect
It isn't clear to me what social annotate does better/faster/simpler.  Please make this clearer in the paper.md

> Substantial scholarly effort: Does this submission meet the scope eligibility described in the JOSS guidelines

https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort

See https://github.com/openjournals/joss-reviews/issues/2780
