### 2020-10-26 uluturki/social-annotate#19 @davekinkead
This will also help satisfy the following requirement in the readme

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
