### 2020-10-26 uluturki/social-annotate#20 @davekinkead
Provide appropriate default settings
The default settings on the twitter-user survey are no-No and yes-Yes. This doesn't make sense (to me at least). Are these key-value pairs or choices yes-no?

ref https://github.com/openjournals/joss-reviews/issues/2780
