### 2020-10-26 uluturki/social-annotate#21 @davekinkead
Clarify state of field
It's not clear to me how social annotate improves on other software in the field.  Please clarify the improvement. See also #19

ref https://github.com/openjournals/joss-reviews/issues/2780
