### 2020-10-26 uluturki/social-annotate#23 @davekinkead
Is the software feature complete?
> In addition, JOSS requires that software should be feature-complete (i.e. no half-baked solutions) and designed for maintainable extension (not one-off modifications of existing tools).

The readme states instagram support and V1.0 is coming soon https://github.com/uluturki/social-annotate#going-public-v10-soon.

I can't see surveys on instagram.

Ref https://github.com/openjournals/joss-reviews/issues/2780
