### 2020-10-26 uluturki/social-annotate#24 @davekinkead
Add automated tests
> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I could not find any automated tests or package.json

Ref https://github.com/openjournals/joss-reviews/issues/2780
