### 2020-10-26 #2780 @davekinkead
Hi @arfon I've completed a first pass and added issues on the software's issue page https://github.com/uluturki/social-annotate/issues.

If the author(s) could let me know when they have  addressed or clarified these, that would be 🥳 ... I'll have another look then.
