### 2020-10-26 #2780 @sara-02
@davekinkead I guess we need an issue for documentation as well, covering the points ` t**he core functionality of the software documented to a satisfactory level (e.g., API method documentation)? and Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)**`. If you have not already covered them in the existing issues I can add one.
 I see I docs folder but I do not think it is connected to a site like readthe docs or something @onurvarol  am I missing something w.r.t doc?
