### 2020-10-27 uluturki/social-annotate#25 @sara-02
[Update Code Documentation and API Examples]
Hey @onurvarol as a part of the JOSS review we also require the docs to be updated for the following

Can you please update the same, or share the portion of the repo where the updated information can be found. Best will be hosting the docs on ReadTheDocs, or if you want to host the APIs then Swagger, so that we have examples of the APIs format redily available.
