### 2020-10-27 #2780 @sara-02
Added: https://github.com/uluturki/social-annotate/issues/25
