+++ manuscript/paper.bib
@@ -17,10 +17,20 @@
 }
 
 @article{potthast2017stylometric,
-  title={A stylometric inquiry into hyperpartisan and fake news},
-  author={Potthast, Martin and Kiesel, Johannes and Reinartz, Kevin and Bevendorff, Janek and Stein, Benno},
-  journal={arXiv preprint arXiv:1702.05638},
-  year={2017}
+  title = "A Stylometric Inquiry into Hyperpartisan and Fake News",
+    author = "Potthast, Martin  and
+      Kiesel, Johannes  and
+      Reinartz, Kevin  and
+      Bevendorff, Janek  and
+      Stein, Benno",
+    booktitle = "Proceedings of the 56th Annual Meeting of the Association for Computational Linguistics (Volume 1: Long Papers)",
+    year = "2018",
+    address = "Melbourne, Australia",
+    publisher = "Association for Computational Linguistics",
+    url = "https://www.aclweb.org/anthology/P18-1022",
+    doi = "10.18653/v1/P18-1022",
+    pages = "231--240",
+    abstract = "We report on a comparative style analysis of hyperpartisan (extremely one-sided) news and fake news. A corpus of 1,627 articles from 9 political publishers, three each from the mainstream, the hyperpartisan left, and the hyperpartisan right, have been fact-checked by professional journalists at BuzzFeed: 97{\%} of the 299 fake news articles identified are also hyperpartisan. We show how a style analysis can distinguish hyperpartisan news from the mainstream (F1 = 0.78), and satire from both (F1 = 0.81). But stylometry is no silver bullet as style-based fake news detection does not work (F1 = 0.46). We further reveal that left-wing and right-wing news share significantly more stylistic similarities than either does with the mainstream. This result is robust: it has been confirmed by three different modeling approaches, one of which employs Unmasking in a novel way. Applications of our results include partisanship detection and pre-screening for semi-automatic fake news detection."
 }
 
 @article{wang2017liar,
@@ -45,7 +55,8 @@
   number={1},
   pages={48--61},
   year={2019},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/hbe2.115}
 }
 
 @inproceedings{varol2017online,
@@ -63,7 +74,8 @@
   number={1},
   pages={1--2},
   year={2010},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/meet.14504701444}
 }
 
 @inproceedings{kaur2014scholarometer,
@@ -72,7 +84,8 @@
   booktitle={Proceedings of the 2014 ACM conference on Web science},
   pages={285--286},
   year={2014},
-  organization={ACM}
+  organization={ACM},
+  doi={10.1371/journal.pone.0043235}
 }
 
 @inproceedings{chakraborty2016stop,
@@ -81,7 +94,8 @@
   booktitle={2016 IEEE/ACM International Conference on Advances in Social Networks Analysis and Mining (ASONAM)},
   pages={9--16},
   year={2016},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/asonam.2016.7752207}
 }
 
 @article{khayat2019vassl,
@@ -92,7 +106,8 @@
   number={1},
   pages={874--883},
   year={2019},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/tvcg.2019.2934266}
 }
 
 @inproceedings{hsueh2009data,
@@ -111,5 +126,6 @@
   number={7},
   pages={e69958},
   year={2013},
-  publisher={Public Library of Science}
+  publisher={Public Library of Science},
+  doi={10.1371/journal.pone.0069958}
 }
