+++ docs/index.md
@@ -2,18 +2,20 @@
 layout: default
 ---
 
-# Content Annotator
+# Social Annotate
 
-Abstract of the paper will go here.
+Human annotated data is essential in building machine learning systems and analyzing online systems. Most research efforts go through cumbersome data collection phases where they either build complicated annotation pipelines for online data or annotate offline data shared on spreadsheets or surveys. These approaches introduce significant overhead to any study requiring data collection and causes researchers to divert significant effort away from analysis and modelling to data collection, especially when they have limited programming experience.
+Here we present `Social-Annotate`, an extendable and configurable browser extension, to assist annotation and collection of online data. In this project we aim to develop a tool that can easily adapt to different tasks and extend to most online platforms. `Social-Annotate` allows collection of data directly on the online platform as a user sees and interacts with it as they normally would by injecting surveys into webpages on the browser, and annotations can easily be exported as files or sent to API endpoints for collaborative data collection. We aim with this tool to promote and expedite annotation of online datasets and reduce the data collection overhead so that researchers can focus on research.
+
 
 ## How to cite 
 
 ```
-@article{contentAnnotator2019,
-  title={},
+@article{socialAnnotate2020,
+  title={Social-Annotate: Browser extension to annotate and collect social media data},
   author={Uluturk, Ismail and Varol, Onur},
-  journal={},
-  year={2019}
+  journal={JOSS (Under review)},
+  year={2020}
 }
 ```
 

[…]
