### 2020-12-14 uluturki/social-annotate#20 @onurvarol
Thank you Dave, we updated the default options and provide more applicable examples.
