### 2020-12-14 uluturki/social-annotate#25 @onurvarol
Dear @sara-02, we updated the README file to provide more details about the software. We also embed a Youtube video with a demo of the tool.
