### 2020-12-14 uluturki/social-annotate#23 @onurvarol
Dear @davekinkead, this tool is extendable to multiple platforms with minimal amount of coding. We updated the Instagram feature as an example, so there are three survey available: twitter-user, twitter-tweet, and instagram-user. Our paper describes how one can extend our tool to other platforms. We also provide a section on how to contribute.
